Make the services more stable
This patch ensures the services do not just give up if one of the others
is missing. They are now going to wait for them to become available and
proceed with the registration and do their work. Listeners are only
started once all required services are found.

Scripts for launching services, workers and uploading image (currently
hardcoded) are also added to simplify deployment.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>