Add a tutorial for the application
This patch adds a complete tutorial on how to use this demo application.
It starts by introducing the application as a whole, then guides you
through running it without the use of OSv. Finally, it explains how to
compose the unikernel image, shows how to run all the services and use
the application transparently.